Recreated the frontend app.
The previous one was created using Microsoft's template, and it wasn't functioning properly. Moreover, it had a lot of strange dependencies.

The new one is created through npx, because it allows to specify the Angular version. 8 is used instead of the latest 9, because ASP.NET Core currently has an issue of not being able to run Angular 9 apps.